T1X: Include cmp-split i cmp
Workaround for blanks: Pretransfer unfortunately turns "+" into " ",
so cmp rules don't write a <b/> after the cmp word. But cmp-split is
not a compound, there's no + to turn into space, so the first <b/>
there is an empty string. To ensure there's a space to output, we put
a space in a variable and use that instead …